requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.2.0
chardet==3.0.4
discord-py-slash-command==1.0.9.5
discord.py==1.5.0
idna==2.10
multidict==5.0.0
redis==3.5.3
yarl==1.6.2
